Tests/reorganize: move wrapText test to own file
